requirements.txt
-e git://github.com/sydney-linux-user-group-dependencies/python-datetime-tz.git@8f081442da51dadee67afacb25901257cfe340be#egg=python_datetime_tz-dev
-e git://github.com/sydney-linux-user-group-dependencies/python-dateutil.git@93f2157fc38500dc73d4f1a4fdbe072be63cd218#egg=python_dateutil-dev
-e git://github.com/sydney-linux-user-group-dependencies/python-markdown.git@5c7355cffd2737de2222cf33fdf047d274af3d44#egg=Markdown-dev
-e git://github.com/sydney-linux-user-group-dependencies/python-pyrss2gen.git@1fde0e73c022e2efc1d51cd655b7369251ace4d7#egg=PyRSS2Gen-dev
-e git://github.com/sydney-linux-user-group-dependencies/python-vobject.git@458fe5d43c7967ff444f48167d04ea6aa9c4a5da#egg=vobject-dev
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django-registration
-e git://github.com/mithro/git-cl.git#egg=git-cl
Django==1.3.1
django-live-server
distribute>=0.6.14
django-extensions==0.7.1
django-social-auth==0.6.1
ez-setup==0.9
httplib2==0.7.2
logilab-astng==0.23.1
logilab-common==0.57.1
oauth2==1.5.211
pylint==0.25.1
python-openid==2.2.5
pytz==2011n
unittest2==0.5.1
wsgiref==0.1.2
gunicorn==0.13.4
selenium
